# WIP: perf: do not fetch all data for workouts to speed up rendering the workout list