When loading the workout list, the app fetches basically the full `workouts` table including all GPX data. This seems a bit unnecessary and also is very slow. This PR fixes it partially by omitting certain fields. However, the `Data` field is still quite large (while the list only uses some statistics). Further work should probaby be done to split the `Data` field into a `Details` field (containing e.g. all the points) and a `Statistics` field containing the basic stats.

This change makes the query a full second faster on my data set (~400 workouts).

Before:

````
12:14PM WRN slow sql query [4.46089625s >= 100ms] app=workout-tracker version=local sha=local module=database slow_query=true query="SELECT * FROM `workouts` WHERE `workouts`.`user_id` = 1 AND `workouts`.`deleted_at` IS NULL ORDER BY date DESC" duration=4.46089625s rows=412 file=/Users/tommy/Repos/workout-tracker/pkg/database/user.go:200
````

After:

````
12:40PM WRN slow sql query [3.434046834s >= 100ms] app=workout-tracker version=local sha=local module=database slow_query=true query="SELECT `workouts`.`id`,`workouts`.`created_at`,`workouts`.`updated_at`,`workouts`.`deleted_at`,`workouts`.`name`,`workouts`.`date`,`workouts`.`user_id`,`workouts`.`dirty`,`workouts`.`notes`,`workouts`.`type`,`workouts`.`data`,`workouts`.`filename` FROM `workouts` WHERE `workouts`.`user_id` = 1 AND `workouts`.`deleted_at` IS NULL ORDER BY date DESC" duration=3.434046834s rows=412 file=/Users/tommy/Repos/workout-tracker/pkg/database/user.go:210
````