@jovandeginste 2024-03-03
I think we should move those other fields to their separate tables. Then it's a matter of (not) preloading those associations.

I was working on this earlier, but didn't get it right yet.