@pixelspark 2024-03-03
Yes, that looks nice :+1: 

It appears `Data` can also get quite heavy, you might want to move that out of the way as well (and only have some basic statistics in the main `workouts` table derived from it).
I am not sure if gorm supports it but you might be able to utilize SQLite JSON functionality (possibly with virtual columns and even indexes on that?) to implement this. (Then again you also want to support other databases)