@jovandeginste 2024-03-03
The plan was to get data out of the way, but that one is a little harder; I wanted to test with the easier column first 😄